+++ media/lua/client/CHC_common.lua
@@ -170,9 +170,11 @@ function CHC_main.common.isRecipeValid(recipe, player, containerList, knownRecip
     else
         if recipe.recipeData.lua.onCanPerform then
             -- print('lua')
-            local luaCanPerformObj = _G[recipe.recipeData.lua.onCanPerform]
-            if luaCanPerformObj then
-                return luaCanPerformObj(recipe, player, nil)
+            local luaCanPerformFunc = recipe.recipeData.lua.onCanPerform
+            if luaCanPerformFunc then
+                return luaCanPerformFunc(recipe, player, nil)
+            else
+                return false
             end
         end
         return true
@@ -517,7 +519,7 @@ function CHC_main.common.getNextState(states, cur)
             end
         end
     end
-    if type(cur) ~= "number" then print("Could not determine current state index") end
+    if type(cur) ~= "number" then utils.chcerror("Could not determine current state index", "getNextState", nil, false) end
     local newStateIx = cur + 1
     if #states < newStateIx then
         newStateIx = 1