+++ media/lua/client/UI/CHC_search.lua
@@ -105,6 +105,7 @@ end
 -- region update
 function CHC_search:update()
     if self.needUpdateDelayedSearch then
+        self.needUpdateDelayedSearch = false
         local props = self.objPanel.itemProps
         props.delayedSearch = CHC_settings.config.delayed_search
         if props.delayedSearch then
@@ -113,6 +114,11 @@ function CHC_search:update()
             props.searchRow:setTooltip(props.searchRow.origTooltip)
         end
     end
+    -- if self.needUpdateObjSource then
+    --     self.needUpdateObjSource = false
+    --     self.objSource = self.backRef[self.objGetter](self, self.ui_type == 'fav_items')
+    --     self:updateObjects()
+    -- end
     CHC_view.update(self)
 end